move state into own package
