+++ pkg/connect/lifecycle.go
@@ -3,6 +3,7 @@ package connect
 import (
 	"context"
 
+	"github.com/inngest/inngest/pkg/connect/state"
 	"github.com/inngest/inngest/proto/gen/connect/v1"
 )
 
@@ -11,7 +12,7 @@ type ConnectGatewayLifecycleListener interface {
 	OnConnected(ctx context.Context, data *connect.SDKConnectRequestData)
 
 	// OnAuthenticated is called when the established connect has successfully authenticated
-	OnAuthenticated(ctx context.Context, auth *AuthResponse)
+	OnAuthenticated(ctx context.Context, auth *state.AuthContext)
 
 	// OnSynced is called when the gateway successfully synced a worker group configuration
 	OnSynced(ctx context.Context)
