+++ pkg/connect/state/error.go
@@ -1,0 +1,12 @@
+package state
+
+import (
+	"fmt"
+)
+
+var (
+	ErrIdempotencyKeyExists = fmt.Errorf("idempotency key exists")
+
+	// internal
+	notImplementedError = fmt.Errorf("not implemented")
+)
